requirements.txt
web3>=7.12,<8
python-dotenv>=1.0.0